# swagger.yaml
openapi: 3.0.3
info:
  title: Photo Studio Booking - API
  version: 1.1.0
  description: >
    API Gateway (5000) for booking & payments + integration endpoints.
    Payment webhook is served by Payment Service (5002).
    Message Broker (5050) и Notification Service (5004) описаны как отдельные servers.
servers:
  - url: http://localhost:5000
    description: API Gateway
  - url: http://localhost:5002
    description: Payment Service (webhooks)
  - url: http://localhost:5050
    description: Message Broker
  - url: http://localhost:5004
    description: Notification Service
paths:
...
  /api/integrations/events:
    get:
      summary: Get integration events log
      tags: [Integrations]
      parameters:
        - in: query
          name: event_type
          required: false
          schema: { type: string, example: payment.succeeded }
        - in: query
          name: source_service
          required: false
          schema: { type: string, example: payment }
        - in: query
          name: limit
          required: false
          schema: { type: integer, default: 100, example: 10 }
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items: { $ref: "#/components/schemas/IntegrationEvent" }
                  total:
                    type: integer
                    example: 3
...
  /api/notifications:
    get:
      summary: Get sent notifications
      tags: [Notifications]
      servers:
        - url: http://localhost:5004
      parameters:
        - in: query
          name: limit
          required: false
          schema: { type: integer, default: 100, example: 20 }
      responses:
        "200":
          description: Notifications list
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifications:
                    type: array
                    items: { $ref: "#/components/schemas/Notification" }
                  total:
                    type: integer
                    example: 5
  /api/notifications/send:
    post:
      summary: Send notification manually
      tags: [Notifications]
      servers:
        - url: http://localhost:5004
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SendNotificationRequest" }
      responses:
        "200":
          description: Notification sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: sent }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Выбранное время недоступно"
...
    IntegrationEvent:
      type: object
      properties:
        event_id: { type: string, format: uuid }
        event_type: { type: string, example: payment.succeeded }
        source_service: { type: string, example: payment }
        payload: { type: object, additionalProperties: true }
        status: { type: string, example: processed }
        timestamp: { type: string, format: date-time }
        processed_at: { type: string, nullable: true, format: date-time }
    Notification:
      type: object
      properties:
        notification_id: { type: string, format: uuid }
        type:
          type: string
          enum: [email, sms]
          example: email
        to: { type: string, example: "ivan@example.com" }
        subject: { type: string, nullable: true, example: "Подтверждение бронирования" }
        body: { type: string, nullable: true, example: "Иван, ваше бронирование подтверждено." }
        message: { type: string, nullable: true, example: "Ваше бронирование подтверждено." }
        status: { type: string, example: sent }
        sent_at: { type: string, format: date-time }
    SendNotificationRequest:
      type: object
      required: [type, to]
      properties:
        type:
          type: string
          enum: [email, sms]
          example: email
        to:
          type: string
          example: "ivan@example.com"
        subject:
          type: string
          nullable: true
          example: "Тема письма"
        body:
          type: string
          nullable: true
          example: "Текст письма"
        message:
          type: string
          nullable: true
          example: "Текст SMS"
...